requirements.txt
numpy
pandas
pyvista
trimesh
torch
h5py
scipy.spatial
pickle
plotly
seaborn
